refactor components path
