+++ src/components/app-header/index.js
@@ -1,0 +1,3 @@
+import AppHeader from './app-header';
+
+export default AppHeader;
