+++ src/components/app/app.css
@@ -6,3 +6,4 @@
 .top-panel {
   margin: 1rem 0;
 }
+
