+++ src/components/item-status-filter/index.js
@@ -1,0 +1,3 @@
+import ItemStatusFilter from './item-status-filter';
+
+export default ItemStatusFilter;
