+++ src/components/search-panel/index.js
@@ -1,0 +1,3 @@
+import SearchPanel from './search-panel';
+
+export default SearchPanel;
