+++ src/components/search-panel/search-panel.css
