+++ src/components/search-panel/search-panel.js
