+++ src/components/todo-list-item/index.js
@@ -1,0 +1,3 @@
+import TodoListItem from './todo-list-item';
+
+export default TodoListItem;
