+++ src/components/todo-list/index.js
@@ -1,0 +1,3 @@
+import TodoList from './todo-list';
+
+export default TodoList;
